familiy polymorphism in scala
This paper was created for a class Multiple paradigm which focused on different
type of programming language features. The assignment was to study the family
polymorphism and create both a paper of the summary and a short presentation for
the class.

This only includes the latex and required files to create the pdf. There is also
additionaly a Makefile in order to build both the slides and the paper itself.
The dependencies are described in the README.md file and mostly consist of latex
softwares, templates.

Signed-off-by: Kokan <kokaipeter@gmail.com>